+++ pkg/home/manager_test.go
@@ -34,9 +34,9 @@ var _ = Describe("home manager", func() {
 					Current: "default",
 					Contexts: []types.Context{
 						{
-							Name:          "default",
-							Builder:       types.BuilderTypeDocker,
-							BuilderSocket: "envd_buildkitd",
+							Name:           "default",
+							Builder:        types.BuilderTypeDocker,
+							BuilderAddress: "envd_buildkitd",
 						},
 					},
 				},
@@ -46,10 +46,11 @@ var _ = Describe("home manager", func() {
 			Expect(m.CacheDir()).To(Equal(filepath.Join(fileutil.DefaultCacheDir)))
 			Expect(m.ConfigFile()).To(Equal(filepath.Join(fileutil.DefaultConfigDir, "config.envd")))
 			Expect(m.ContextFile()).To(Equal(filepath.Join(fileutil.DefaultConfigDir, "contexts")))
-			driver, socket, err := m.ContextGetCurrent()
+			c, err := m.ContextGetCurrent()
 			Expect(err).NotTo(HaveOccurred())
-			Expect(driver).To(Equal(types.BuilderTypeDocker))
-			Expect(socket).To(Equal("envd_buildkitd"))
+			Expect(c.Builder).To(Equal(types.BuilderTypeDocker))
+			Expect(c.BuilderAddress).To(Equal("envd_buildkitd"))
+			Expect(c.Runner).To(Equal(types.RunnerTypeDocker))
 		})
 		It("should return the cache status", func() {
 			Expect(os.RemoveAll(filepath.Join(fileutil.DefaultCacheDir, "cache.status"))).NotTo(HaveOccurred())