+++ pkg/lang/ir/fs.go
@@ -11,14 +11,15 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 package ir
 
 import (
 	"github.com/moby/buildkit/client/llb"
 )
 
-func (g Graph) CompileCacheDir(root llb.State, cacheDir string) llb.State {
-	root = llb.User("envd")(root)
+func (g *Graph) CompileCacheDir(root llb.State, cacheDir string) llb.State {
+	g.UserDirectories = append(g.UserDirectories, cacheDir)
 	run := root.Run(llb.Shlexf("mkdir -p %s", cacheDir), llb.WithCustomName("[internal] create cache dir"))
 	return run.Root()
 }