+++ pkg/lang/ir/python.go
@@ -24,7 +24,6 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/tensorchord/envd/pkg/flag"
-	"github.com/tensorchord/envd/pkg/util/fileutil"
 )
 
 const (
@@ -105,7 +104,6 @@ func (g Graph) compilePython(aptStage llb.State) (llb.State, error) {
 
 // Set the system default python to envd's python.
 func (g Graph) compileAlternative(root llb.State) llb.State {
-	root = llb.User("root")(root)
 	envdPrefix := "/opt/conda/envs/envd/bin"
 	run := root.
 		Run(llb.Shlexf("update-alternatives --install /usr/bin/python python %s/python 1", envdPrefix), llb.WithCustomName("update alternative python to envd")).
@@ -120,12 +118,12 @@ func (g Graph) compilePyPIPackages(root llb.State) llb.State {
 		return root
 	}
 
-	cacheDir := fileutil.EnvdHomeDir(".cache")
-	// Create the cache directory to the container. see issue #582
+	// Create the envd cache directory in the container. see issue #582
+	cacheDir := filepath.Join("/", "root", ".cache", "pip")
 	root = g.CompileCacheDir(root, cacheDir)
 
-	cache := root.File(llb.Mkdir("/cache",
-		0755, llb.WithParents(true), llb.WithUIDGID(g.uid, g.gid)), llb.WithCustomName("[internal] setting pip cache mount permissions"))
+	cache := root.File(llb.Mkdir("/cache/pip", 0755, llb.WithParents(true)),
+		llb.WithCustomName("[internal] setting pip cache mount permissions"))
 
 	if len(g.PyPIPackages) != 0 {
 		// Compose the package install command.
@@ -139,13 +137,12 @@ func (g Graph) compilePyPIPackages(root llb.State) llb.State {
 		cmd := sb.String()
 		logrus.WithField("command", cmd).
 			Debug("Configure pip install statements")
-		root = llb.User("envd")(root)
 		run := root.
 			Run(llb.Shlex(sb.String()), llb.WithCustomNamef("pip install %s",
 				strings.Join(g.PyPIPackages, " ")))
 		// Refer to https://github.com/moby/buildkit/blob/31054718bf775bf32d1376fe1f3611985f837584/frontend/dockerfile/dockerfile2llb/convert_runmount.go#L46
 		run.AddMount(cacheDir, cache,
-			llb.AsPersistentCacheDir(g.CacheID(cacheDir), llb.CacheMountShared), llb.SourcePath("/cache"))
+			llb.AsPersistentCacheDir(g.CacheID(cacheDir), llb.CacheMountShared), llb.SourcePath("/cache/pip"))
 		root = run.Root()
 	}
 
@@ -170,7 +167,7 @@ func (g Graph) compilePyPIPackages(root llb.State) llb.State {
 		run := root.
 			Run(llb.Shlex(cmd), llb.WithCustomNamef("pip install %s", *g.RequirementsFile))
 		run.AddMount(cacheDir, cache,
-			llb.AsPersistentCacheDir(g.CacheID(cacheDir), llb.CacheMountShared), llb.SourcePath("/cache"))
+			llb.AsPersistentCacheDir(g.CacheID(cacheDir), llb.CacheMountShared), llb.SourcePath("/cache/pip"))
 		run.AddMount(g.getWorkingDir(),
 			llb.Local(flag.FlagBuildContext))
 		root = run.Root()
@@ -183,7 +180,7 @@ func (g Graph) compilePyPIPackages(root llb.State) llb.State {
 			run := root.Run(llb.Shlex(fmt.Sprintf(cmdTemplate, wheel)), llb.WithCustomNamef("pip install %s", wheel))
 			run.AddMount(g.getWorkingDir(), llb.Local(flag.FlagBuildContext), llb.Readonly)
 			run.AddMount(cacheDir, cache,
-				llb.AsPersistentCacheDir(g.CacheID(cacheDir), llb.CacheMountShared), llb.SourcePath("/cache"))
+				llb.AsPersistentCacheDir(g.CacheID(cacheDir), llb.CacheMountShared), llb.SourcePath("/cache/pip"))
 			root = run.Root()
 		}
 	}